% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{calculate_mean_of_mean}
\alias{calculate_mean_of_mean}
\title{Calculate Mean of Mean Uniqueness}
\usage{
calculate_mean_of_mean(data)
}
\arguments{
\item{data}{A data frame containing the mean uniqueness scores.}
}
\value{
A single numeric value representing the mean of the mean uniqueness scores.
}
\description{
This function calculates the mean of the mean uniqueness scores.
}
