% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{calculate_name_frequency}
\alias{calculate_name_frequency}
\title{Calculate Name Frequency}
\usage{
calculate_name_frequency(data)
}
\arguments{
\item{data}{A data frame containing the data with 'sample' columns.}
}
\value{
A data frame with the frequency of each name.
}
\description{
This function calculates the frequency of names in the provided data.
}
