% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{calculate_uniqueness}
\alias{calculate_uniqueness}
\title{Calculate Uniqueness Score}
\usage{
calculate_uniqueness(data)
}
\arguments{
\item{data}{A data frame containing the frequency of names.}
}
\value{
A data frame with the uniqueness score for each name.
}
\description{
This function calculates the uniqueness score for each name in the data.
}
