% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{create_year_windows}
\alias{create_year_windows}
\title{Create Year Windows}
\usage{
create_year_windows(df, window_size)
}
\arguments{
\item{df}{A data frame containing a 'year' column.}

\item{window_size}{An integer specifying the size of the year window.}
}
\value{
A data frame with an additional 'year_window' column indicating the year window for each row.
}
\description{
This function creates year windows by grouping years into specified intervals.
}
